+++ EventManager.js
@@ -1,0 +1,63 @@
+class EventManager {
+    constructor() {
+    }
+    /**
+     * Deletes an event.
+     */
+    delete() {
+    }
+    /**
+     * Returns an event.
+     */
+    get() {
+    }
+    /**
+     * Imports an event this operation is used to add a private copy of an
+     * existing event to a calendar.
+     */
+    import() {
+    }
+    /**
+     * Creates an event
+     */
+    insert() {
+    }
+    /**
+     * Returns instances of the specified recurring event.
+     */
+    instances() {
+    }
+    /**
+     * Returns events on the specified calendar.
+     */
+    list() {
+    }
+    /**
+     * Moves an event to another calendar, i.e, changes an event's organizer.
+     */
+    move() {
+    }
+    /**
+     * Updates an event. This method supports patch semantics. The field values
+     * you specify replace the existing values. Fields that you don't specify
+     * in the request remain unchanged. Array fields, if specified, overwrite the
+     * existing arrays; this discards any previous array elements.
+     */
+    patch() {
+    }
+    /**
+     * Creates an event based on a simple text string
+     */
+    quickAdd() {
+    }
+    /**
+     * Updates an event
+     */
+    update() {
+    }
+    /**
+     * Watch for changes to Events resources
+     */
+    watch() {
+    }
+}
