+++ EventManager.ts
@@ -1,0 +1,86 @@
+class EventManager {
+    constructor() {
+
+    }
+
+    /**
+     * Deletes an event.
+     */
+    public delete(): void {
+
+    }
+
+    /**
+     * Returns an event.
+     */
+    public get(): void {
+
+    }
+
+    /**
+     * Imports an event this operation is used to add a private copy of an
+     * existing event to a calendar.
+     */
+    public import(): void {
+
+    }
+
+    /**
+     * Creates an event
+     */
+    public insert(): void {
+
+    }
+
+    /**
+     * Returns instances of the specified recurring event.
+     */
+    public instances(): void {
+
+    }
+
+    /**
+     * Returns events on the specified calendar.
+     */
+    public list(): void {
+
+    }
+
+    /**
+     * Moves an event to another calendar, i.e, changes an event's organizer.
+     */
+    public move(): void {
+
+    }
+
+    /**
+     * Updates an event. This method supports patch semantics. The field values
+     * you specify replace the existing values. Fields that you don't specify
+     * in the request remain unchanged. Array fields, if specified, overwrite the
+     * existing arrays; this discards any previous array elements.
+     */
+    public patch(): void {
+
+    }
+
+    /**
+     * Creates an event based on a simple text string
+     */
+    public quickAdd(): void {
+
+    }
+
+    /**
+     * Updates an event
+     */
+    public update(): void {
+
+    }
+
+    /**
+     * Watch for changes to Events resources
+     */
+    public watch(): void {
+
+    }
+}
