+++ GoogleCalendarApi.tests.ts
@@ -1,0 +1,8 @@
+declare var test: any;
+
+import GoogleCalendarApi from './build/GoogleCalendarApi';
+
+test('setCalendar method', () => {
+    GoogleCalendarApi.setCalendar('test-calendar');
+    expect(GoogleCalendarApi.calendar).toBe('test-calendar');
+});
